Use ServiceLoader to load tiles.
