+++ integrations/beast3/java/src/main/java/PhyloSpecRunner.java
@@ -19,8 +19,9 @@ import org.phylospec.typeresolver.TypeError;
 import org.phylospec.typeresolver.TypeResolver;
 import org.phylospec.typeresolver.VariableResolver;
 import org.xml.sax.SAXException;
-import tiles.TileLibrary;
+import tiles.BeastCoreTileLibrary;
 import beastconfig.BEASTState;
+import tiles.TileLibrary;
 import tiling.EvaluateTiles;
 import tiling.TileApplicationError;
 
@@ -91,7 +92,7 @@ public class PhyloSpecRunner implements ErrorEventListener {
 
         // perform tiling
 
-        EvaluateTiles applyTiles = new EvaluateTiles(TileLibrary.getTiles(), OperatorTileLibrary.getTiles(), variableResolver, stochasticityResolver);
+        EvaluateTiles applyTiles = new EvaluateTiles(TileLibrary.loadAll(), OperatorTileLibrary.getTiles(), variableResolver, stochasticityResolver);
         BEASTState beastState = new BEASTState(runName);
         try {
             applyTiles.getBestTiling(statements);
