+++ integrations/beast3/java/src/main/resources/META-INF/services/tiles.TileLibrary
@@ -1,0 +1,1 @@
+tiles.BeastCoreTileLibrary
